fix(ci): python-coverage-shards — split 1601-line drift test under 1500-line maintainability limit
test_first_party_code_files_stay_under_line_limit failed because the new
tests/unit/docs/test_command_grammar_drift.py was 1601 lines (limit 1500).
Split it into a command_grammar_drift_parts/ package following the existing
test_*_parts convention: shared constants + parsing helpers in _helpers.py,
helper-level tests in part_001, and the real-doc R1-R5 sweeps in part_002.
No behavior change — same 24 tests, REPO_ROOT adjusted for the extra dir level.
